+++ apps/eaas-interface/src/api/embalm.ts
@@ -4,7 +4,10 @@ import { handleApiError } from "./utils";
 
 export async function sendPayload(params: SendEncryptedPayloadParams) {
   try {
-    const res = await axios.post(`${import.meta.env.VITE_API_BASE_URL}/embalm/send-payload`, params);
+    const res = await axios.post(
+      `${import.meta.env.VITE_API_BASE_URL}/embalm/send-payload`,
+      params,
+    );
     return res.status === 200;
   } catch (error) {
     handleApiError(error);
